Refactor router out into two separate routers
The current router was starting to get quite long therefore harder
to maintain. It also included lots of duplication in the route
definitions with the public token param.

This refactor moves all the logic reliant on the public token path
into its own router and making it easier to maintain.