Create a payment gateway session
This change creats a payment gateway session to redirect the
payment journey to on a `GET` of the card payment step.

It also updates the card view to simply include a link that is
now set within a piece of middleware rather than containing a
form that posts back to itself.