Revert package lock file
During the addition of the client-sessions dependency some how the
package lock file got a bit corrupted and dropped some dependencies
that were required.

This change reset the package lock file to a previous version without
this dependency and with a fresh `npm install`.

This removes the error with the missing sub-dependency but the server
does fail to start because it is missing client-sessions.

This will be re-added in the following commit.