Only store payment gateway session ID in session
This refactor only stores the payment gateway session ID in the
users session rather than the whole object. It then checks the API
again using this ID to get the latest state and payment URL of this
payment session.

This will also be used in following changes to checkt the status
of the payment session and handle it appropriately for possible
edge cases.