Ensure payment response routes aren't prevented
The paid status middleware renders an alternative template if the
order has been paid to prevent users entering the payment journey
again.

However, this prevented the success/failure and return routes being
accessed because by that time, after a successful payment the API
will have changed the order state to paid.

This change re-orders the middleware so that checks only apply as and
when they need to.