Handle conflicted payment gateway session
If the creation of a payment gateway returns a conflict, or `409`
at this point it will mean that the payment will have been successful
but may not have been marked as paid.

The way we need to handle it is to redirect back to the payment screen
as calling this endpoint will reconcile the payment and change the
order status to paid.