Display the full payment reference
This uses the payment reference from the payments endpoint rather
than just the order reference number on the payment confirmation page.

This is so that the reference number matches the one sent to the
user in an email and also so it will be easier for the Live Service
Team to debug and possible problems.