Update logic and prominence of quote time left
This change updates the logic and the messaging for the different
states of a quote.

It reduces the prominence of the message when a quote still has time
left and uses the metalist style rather than a large banner.

It also removes the banner for accepted quotes in place of the
metalist style.